Move meta_index to vast::system
