+++ libvast/vast/fwd.hpp
@@ -75,7 +75,6 @@ class data;
 class ewah_bitstream;
 class expression;
 class json;
-class meta_index;
 class msgpack_table_slice_builder;
 class path;
 class pattern;
@@ -235,6 +234,7 @@ class application;
 class configuration;
 class default_application;
 class export_command;
+class meta_index;
 class node_command;
 class pcap_writer_command;
 class remote_command;
